clean up folder structure, separate social media icons into its own component
