+++ vue/src/components/body/VueFooter.vue
@@ -1,13 +1,13 @@
 <template>
     <footer>
-		<h6 v-html="message"></h6>
+		<h6 v-html="footerText"></h6>
 	</footer>
 </template>
 
 <script>
 export default {
     computed: {
-        message: function() {
+        footerText: function() {
             const year = new Date().getFullYear();
             return "&copy; Tyler Earls " + year;
         }
@@ -15,8 +15,9 @@ export default {
 }
 </script>
 
-<style scoped>
-    h6 {
-        text-align: center;
-    }
+<style lang="scss" scoped>
+
+h6 {
+    text-align: center;
+}
 </style>
