+++ vue/src/components/body/WebProject.vue
@@ -47,39 +47,40 @@ export default {
 </script>
 
 <style lang="scss" scoped>
-    @import 'src/scss/global.scss';
-    h1 {
-        margin-bottom: 20px;
-    }
-    .portfolio-link {
-        display: block;
-        color: $red;
-        font-weight: bold;
-        text-align: center;
-        cursor: pointer;
-        margin-bottom: 5px;
-    }
-    .portfolio-screenshot {
-        margin: 7px auto;
-        max-width: calc(100vw - #{$navWidth} - 40px);
-        height: auto;
-        display: block;
-    }
-    .project-title, .skill-title {
-        margin-bottom: 10px;
-        font-weight: bold;
-        color: $red;
-        text-align: center;
-    }
-    .project-description {
-        margin-bottom: 10px;
-    }
-    .line-break {
-        margin-top: 15px;
-        border: 0;
-        height: 1px;
-        margin-left: 11vw;
-        margin-right: 11vw;
-        background: linear-gradient(to right, rgba(0, 0, 0, 0), $red, rgba(0, 0, 0, 0));
-    }
+@import 'src/scss/global.scss';
+
+h1 {
+    margin-bottom: 20px;
+}
+.portfolio-link {
+    display: block;
+    color: $red;
+    font-weight: bold;
+    text-align: center;
+    cursor: pointer;
+    margin-bottom: 5px;
+}
+.portfolio-screenshot {
+    margin: 7px auto;
+    max-width: calc(100vw - #{$navWidth} - 40px);
+    height: auto;
+    display: block;
+}
+.project-title, .skill-title {
+    margin-bottom: 10px;
+    font-weight: bold;
+    color: $red;
+    text-align: center;
+}
+.project-description {
+    margin-bottom: 10px;
+}
+.line-break {
+    margin-top: 15px;
+    border: 0;
+    height: 1px;
+    margin-left: 11vw;
+    margin-right: 11vw;
+    background: linear-gradient(to right, rgba(0, 0, 0, 0), $red, rgba(0, 0, 0, 0));
+}
 </style>
