+++ vue/src/components/renderless/ErrorMessage.vue
@@ -17,32 +17,32 @@ export default {
 </script>
 
 <style scoped lang="scss">
-    @import 'src/scss/global.scss';
+@import 'src/scss/global.scss';
 
-    .error-placeholder {
-        display: block;
-        height: 20px;
-    }
-    .success-message {
-        font-weight: 400;
-        color: $red;
-        -webkit-font-smoothing: antialiased;
-        -moz-osx-font-smoothing: grayscale;
-        font-size: 13px;
-        width: fit-content;
-    }
-    .error-message {
-        display: block;
-        font-weight: 400;
-        color: $red;
-        -webkit-font-smoothing: antialiased;
-        -moz-osx-font-smoothing: grayscale;
-        border-radius: 0.25em;
-        width: fit-content;
-        font-size: 13px;
-        height: 20px;
-        margin-left: initial;
-        margin-right: initial;
-        font-weight: 400;
-    }
+.error-placeholder {
+    display: block;
+    height: 20px;
+}
+.success-message {
+    font-weight: 400;
+    color: $red;
+    -webkit-font-smoothing: antialiased;
+    -moz-osx-font-smoothing: grayscale;
+    font-size: 13px;
+    width: fit-content;
+}
+.error-message {
+    display: block;
+    font-weight: 400;
+    color: $red;
+    -webkit-font-smoothing: antialiased;
+    -moz-osx-font-smoothing: grayscale;
+    border-radius: 0.25em;
+    width: fit-content;
+    font-size: 13px;
+    height: 20px;
+    margin-left: initial;
+    margin-right: initial;
+    font-weight: 400;
+}
 </style>
